+++ CHANGES.rst
@@ -8,6 +8,8 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^^
 
 - Added __hash__ method to Country class
+- Made Country validate itself during object initialization
+- Made Country string coercible
 
 
 0.29.9 (2015-04-07)